split background logic, added comments
